# docs/openapi.yaml
openapi: 3.1.0
info:
  title: InkFlow CRM API
  description: |
    REST API for InkFlow tattoo studio platform.

    Base path: /api (server context-path).

    Response envelope: ApiResponse<T> - see docs/API_FORMAT.md.

    Auth:
    - CRM (admin mobile / web): Supabase JWT in Authorization Bearer header.
      Custom claims: tenant_id, role, location_ids (via Supabase hook).
    - Consumer (B2C app): Supabase JWT for /public/consumer/** routes.
    - Public reads: GET /public/catalog/** and GET /public/artists/** - no token.
  contact:
    name: InkFlow
    email: dev@inkflow.studio
  license:
    name: Proprietary
  version: 1.0.0
servers:
- url: /api
  description: Current host (context-path included)
paths:
...
  /staff/{id}/services/{serviceId}:
    put:
      tags:
      - CRM Â· Staff
      operationId: updateStaffServicePricing
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStaffServiceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffServiceDto"
    post:
      tags:
      - CRM Â· Staff
      operationId: addServiceToStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStaffServiceRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseStaffServiceDto"
    delete:
      tags:
      - CRM Â· Staff
      operationId: removeServiceFromStaff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: serviceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /services:
    get:
      tags:
      - CRM Â· Services
      operationId: getAllServices
      parameters:
      - name: pageRequest
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/PageRequest"
      - name: active
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListServiceDto"
    post:
      tags:
      - CRM Â· Services
      operationId: createService
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseServiceDto"
...
  /services/{id}:
    get:
      tags:
      - CRM Â· Services
      operationId: getService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseServiceDetailDto"
    delete:
      tags:
      - CRM Â· Services
      operationId: deleteService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
    patch:
      tags:
      - CRM Â· Services
      operationId: updateService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseServiceDto"
...
  /services/{id}/price:
    get:
      tags:
      - CRM Â· Services
      operationId: getServicePrice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: artistId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseServicePriceDto"
...
components:
  schemas:
...
    ErrorDto:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorDto"
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    PaginationDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    StaffServiceDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        pricingType:
          type: string
        basePrice:
          type: number
        customPrice:
          type: number
        baseDuration:
          type: integer
          format: int32
        customDuration:
          type: integer
          format: int32
        color:
          type: string
        isActive:
          type: boolean
    AddStaffServiceRequest:
      type: object
      properties:
        customPrice:
          type: number
        customDuration:
          type: integer
          format: int32
    ApiResponseStaffServiceDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/StaffServiceDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
...
    CreateServiceRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 2
        description:
          type: string
        pricingType:
          type: string
          pattern: ^(fixed|hourly|project)$
        price:
          type: number
          minimum: 0.0
        duration:
          type: integer
          format: int32
          minimum: 15
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        costPrice:
          type: number
          minimum: 0.0
      required:
      - color
      - duration
      - price
      - pricingType
      - title
    ApiResponseServiceDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ServiceDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ServiceDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        pricingType:
          type: string
        price:
          type: number
        duration:
          type: integer
          format: int32
        color:
          type: string
        isActive:
          type: boolean
        costPrice:
          type: number
        createdAt:
          type: string
          format: date-time
...
    UpdateServiceRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 2
        description:
          type: string
        pricingType:
          type: string
          pattern: ^(fixed|hourly|project)$
        price:
          type: number
          minimum: 0.0
        duration:
          type: integer
          format: int32
          minimum: 15
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
        isActive:
          type: boolean
        costPrice:
          type: number
          minimum: 0.0
...
    PageRequest:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        sortBy:
          type: string
        sortDir:
          type: string
...
    ApiResponseListServiceDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServiceDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ApiResponseServiceDetailDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ServiceDetailDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ArtistPricingOverrideDto:
      type: object
      properties:
        artistId:
          type: string
          format: uuid
        artistName:
          type: string
        price:
          type: number
        duration:
          type: integer
          format: int32
    ServiceDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        pricingType:
          type: string
        price:
          type: number
        duration:
          type: integer
          format: int32
        color:
          type: string
        isActive:
          type: boolean
        costPrice:
          type: number
        artistPricing:
          type: array
          items:
            $ref: "#/components/schemas/ArtistPricingOverrideDto"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApiResponseServicePriceDto:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ServicePriceDto"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ServicePriceDto:
      type: object
      properties:
        serviceId:
          type: string
          format: uuid
        artistId:
          type: string
          format: uuid
        price:
          type: number
        duration:
          type: integer
          format: int32
        isOverride:
          type: boolean
...
  securitySchemes:
    CrmBearer:
      type: http
      description: Supabase JWT for CRM staff (OWNER/ADMIN/ARTIST).
      scheme: bearer
      bearerFormat: JWT
    ConsumerBearer:
      type: http
      description: Supabase JWT for B2C consumer app users.
      scheme: bearer
      bearerFormat: JWT